fix ExtractTextures
